+++ src/textual/binding.py
@@ -50,6 +50,8 @@ class Binding:
     """How the key should be shown in footer."""
     priority: bool = False
     """Enable priority binding for this key."""
+    tooltip: str = ""
+    """Optional tooltip to show in footer."""
 
 
 class ActiveBinding(NamedTuple):
@@ -61,6 +63,8 @@ class ActiveBinding(NamedTuple):
     """The binding information."""
     enabled: bool
     """Is the binding enabled? (enabled bindings are typically rendered dim)"""
+    tooltip: str = ""
+    """Optional tooltip shown in Footer."""
 
 
 @rich.repr.auto
@@ -112,6 +116,7 @@ class BindingsMap:
                         show=bool(binding.description and binding.show),
                         key_display=binding.key_display,
                         priority=binding.priority,
+                        tooltip=binding.tooltip,
                     )
 
         self.key_to_bindings: dict[str, list[Binding]] = {}